HW-3 rencent dialog && adapter package position move
